awaitfs: fix signed int/long error

+++ include/syscall.def.h
@@ -155,7 +155,7 @@ int sigreturn(void);
 int sigprocmask(int how, unsigned long set, unsigned long *old_set);
 
 /// num=0x80
-int awaitfs(struct await_target *, int nfds, int flags, unsigned long timeout_time);
+int awaitfs(struct await_target *, int nfds, int flags, long long timeout_time);
 
 /// num=0xF0
 unsigned long kshell(char *cmd, int argc, char **argv, void *data, size_t len);
